+++ startpage.config.js
@@ -97,170 +97,172 @@ const defaultConfig = {
 			}
 		]
 	},
-	sections: [
-		{
-			title: "General",
-			color: "green",
-			align: "left",
-			links: [
-				{
-					name: "Portfolio",
-					url: "https://cancellek.com",
-					icon: "mdi:web"
-				},
-				{
-					name: "Keybase",
-					url: "https://keybase.io/",
-					icon: "fa-brands:keybase"
-				},
-				{
-					name: "GPT",
-					url: "https://chat.openai.com/",
-					icon: "simple-icons:openai"
-				},
-				{
-					name: "OCI",
-					url: "https://www.oracle.com/cloud/",
-					icon: "simple-icons:oracle"
-				}
-			]
-		},
-		{
-			title: "Dev",
-			color: "magenta",
-			align: "left",
-			links: [
-				{
-					name: "GitHub",
-					url: "https://github.com",
-					icon: "mdi:github"
-				},
-				{
-					name: "GitLab",
-					url: "https://gitlab.com",
-					icon: "ph:gitlab-logo-simple-fill"
-				},
-				{
-					name: "Dev.to",
-					url: "https://dev.to",
-					icon: "material-symbols:logo-dev"
-				},
-				{
-					name: "Stack Overflow",
-					url: "https://stackoverflow.com/",
-					icon: "mdi:stack-overflow"
-				}
-			]
-		},
-		{
-			title: "Social",
-			color: "violet",
-			align: "left",
-			links: [
-				{
-					name: "Twitter",
-					url: "https://twitter.com",
-					icon: "mdi:twitter"
-				},
-				{
-					name: "Mastodon",
-					url: "https://mastodon.social/",
-					icon: "ri:mastodon-fill"
-				},
-				{
-					name: "Reddit",
-					url: "https://reddit.com",
-					icon: "mdi:reddit"
-				},
-				{
-					name: "Polywork",
-					url: "https://polywork.com",
-					icon: "simple-icons:polywork"
-				}
-			]
-		},
-		{
-			title: "Gaming",
-			color: "cyan",
-			align: "left",
-			links: [
-				{
-					name: "Polygon",
-					url: "https://polygon.com",
-					icon: "uil:polygon"
-				},
-				{
-					name: "IGN",
-					url: "https://ign.com",
-					icon: "mdi:currency-sign"
-				},
-				{
-					name: "RPS",
-					url: "https://rockpapershotgun.com/",
-					icon: "ph:toilet-paper-bold"
-				},
-				{
-					name: "80lv",
-					url: "https://80.lv/",
-					icon: "tabler:hand-rock"
-				}
-			]
-		},
-		{
-			title: "Science",
-			color: "blue",
-			align: "left",
-			links: [
-				{
-					name: "PopSci",
-					url: "https://popsci.com/",
-					icon: "material-symbols:science"
-				},
-				{
-					name: "Space",
-					url: "fa6-solid:user-astronaut",
-					icon: "mdi:reddit"
-				},
-				{
-					name: "NASA",
-					url: "https://blogs.nasa.gov/",
-					icon: "simple-icons:nasa"
-				},
-				{
-					name: "ESA",
-					url: "https://blogs.esa.int/",
-					icon: "mdi:black-mesa"
-				}
-			]
-		},
-		{
-			title: "Tech",
-			color: "yellow",
-			align: "left",
-			links: [
-				{
-					name: "TechCrunch",
-					url: "https://techcrunch.com/",
-					icon: "game-icons:techno-heart"
-				},
-				{
-					name: "Verge",
-					url: "https://www.theverge.com/",
-					icon: "arcticons:verge"
-				},
-				{
-					name: "It's Foss",
-					url: "https://itsfoss.com/",
-					icon: "ri:mastodon-fill"
-				},
-				{
-					name: "9To5 Linux",
-					url: "https://9to5linux.com/",
-					icon: "uil:linux"
-				}
-			]
-		}
-	]
+	sections: {
+		list: [
+			{
+				title: "General",
+				color: "green",
+				align: "left",
+				links: [
+					{
+						name: "Portfolio",
+						url: "https://cancellek.com",
+						icon: "mdi:web"
+					},
+					{
+						name: "Keybase",
+						url: "https://keybase.io/",
+						icon: "fa-brands:keybase"
+					},
+					{
+						name: "GPT",
+						url: "https://chat.openai.com/",
+						icon: "simple-icons:openai"
+					},
+					{
+						name: "OCI",
+						url: "https://www.oracle.com/cloud/",
+						icon: "simple-icons:oracle"
+					}
+				]
+			},
+			{
+				title: "Dev",
+				color: "magenta",
+				align: "left",
+				links: [
+					{
+						name: "GitHub",
+						url: "https://github.com",
+						icon: "mdi:github"
+					},
+					{
+						name: "GitLab",
+						url: "https://gitlab.com",
+						icon: "ph:gitlab-logo-simple-fill"
+					},
+					{
+						name: "Dev.to",
+						url: "https://dev.to",
+						icon: "material-symbols:logo-dev"
+					},
+					{
+						name: "Stack Overflow",
+						url: "https://stackoverflow.com/",
+						icon: "mdi:stack-overflow"
+					}
+				]
+			},
+			{
+				title: "Social",
+				color: "violet",
+				align: "left",
+				links: [
+					{
+						name: "Twitter",
+						url: "https://twitter.com",
+						icon: "mdi:twitter"
+					},
+					{
+						name: "Mastodon",
+						url: "https://mastodon.social/",
+						icon: "ri:mastodon-fill"
+					},
+					{
+						name: "Reddit",
+						url: "https://reddit.com",
+						icon: "mdi:reddit"
+					},
+					{
+						name: "Polywork",
+						url: "https://polywork.com",
+						icon: "simple-icons:polywork"
+					}
+				]
+			},
+			{
+				title: "Gaming",
+				color: "cyan",
+				align: "left",
+				links: [
+					{
+						name: "Polygon",
+						url: "https://polygon.com",
+						icon: "uil:polygon"
+					},
+					{
+						name: "IGN",
+						url: "https://ign.com",
+						icon: "mdi:currency-sign"
+					},
+					{
+						name: "RPS",
+						url: "https://rockpapershotgun.com/",
+						icon: "ph:toilet-paper-bold"
+					},
+					{
+						name: "80lv",
+						url: "https://80.lv/",
+						icon: "tabler:hand-rock"
+					}
+				]
+			},
+			{
+				title: "Science",
+				color: "blue",
+				align: "left",
+				links: [
+					{
+						name: "PopSci",
+						url: "https://popsci.com/",
+						icon: "material-symbols:science"
+					},
+					{
+						name: "Space",
+						url: "fa6-solid:user-astronaut",
+						icon: "mdi:reddit"
+					},
+					{
+						name: "NASA",
+						url: "https://blogs.nasa.gov/",
+						icon: "simple-icons:nasa"
+					},
+					{
+						name: "ESA",
+						url: "https://blogs.esa.int/",
+						icon: "mdi:black-mesa"
+					}
+				]
+			},
+			{
+				title: "Tech",
+				color: "yellow",
+				align: "left",
+				links: [
+					{
+						name: "TechCrunch",
+						url: "https://techcrunch.com/",
+						icon: "game-icons:techno-heart"
+					},
+					{
+						name: "Verge",
+						url: "https://www.theverge.com/",
+						icon: "arcticons:verge"
+					},
+					{
+						name: "It's Foss",
+						url: "https://itsfoss.com/",
+						icon: "ri:mastodon-fill"
+					},
+					{
+						name: "9To5 Linux",
+						url: "https://9to5linux.com/",
+						icon: "uil:linux"
+					}
+				]
+			}
+		]
+	}
 }
 
 export default defaultConfig
